requirements.txt
pytz
python-dateutil
prometheus_client
docker